refactor(volunteerApplicationLogic): extract application status rules into pure logic
